# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tool Control Plane API
  version: 0.1.0
  description: Governed tool gateway for AI agents with policy, approvals, audit, and provider adapters.
servers:
  - url: http://localhost:4100
security:
  - bearerAuth: []
paths:
...
  /mcp:
    post:
      summary: MCP JSON-RPC endpoint
      description: Supports initialize, tools/list, tools/call, resources/list, and resources/read over JSON-RPC 2.0.
      operationId: callMCP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPJSONRPCRequest"
      responses:
        "200":
          description: MCP JSON-RPC response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPJSONRPCResponse"
        "204":
          description: Notification accepted with no response body.
...
  /v1/connectors:
    get:
      summary: List configured and registered connectors
      operationId: listConnectors
      responses:
        "200":
          description: Connector inventory without raw secret values.
          content:
            application/json:
              schema:
                type: object
                required: [connectors]
                properties:
                  connectors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Connector"
    post:
      summary: Register connector metadata
      description: Stores non-secret connector metadata for internal alpha. It does not hot-swap provider adapters in the current process.
      operationId: createConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorCreateRequest"
      responses:
        "201":
          description: Connector metadata was registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "400":
          description: Invalid connector request.
...
  /v1/tool-calls:
    get:
      summary: List stored tool call records
      operationId: listToolCalls
      responses:
        "200":
          description: Ordered tool call records with redacted arguments and results.
          content:
            application/json:
              schema:
                type: object
                required: [tool_calls]
                properties:
                  tool_calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/ToolCallRecord"
    post:
      summary: Execute or request approval for a governed tool call
      operationId: callTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCallRequest"
      responses:
        "200":
          description: Tool call result, policy denial, validation error, or approval request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallResponse"
        "400":
          description: Invalid JSON request body.
  /v1/tool-calls/{id}:
    get:
      summary: Get one stored tool call record
      operationId: getToolCall
      parameters:
        - $ref: "#/components/parameters/ToolCallID"
      responses:
        "200":
          description: Tool call record with redacted arguments and results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallRecord"
        "404":
          description: Tool call record was not found.
...
  /v1/approvals/{id}:
    get:
      summary: Get one approval request
      operationId: getApproval
      parameters:
        - $ref: "#/components/parameters/ApprovalID"
      responses:
        "200":
          description: Approval request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
        "404":
          description: Approval was not found.
  /v1/approvals/{id}/grant:
    post:
      summary: Grant a pending approval request
      operationId: grantApproval
      parameters:
        - $ref: "#/components/parameters/ApprovalID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Updated approval decision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "400":
          description: Invalid JSON request body.
        "404":
          description: Approval was not found.
  /v1/approvals/{id}/deny:
    post:
      summary: Deny a pending approval request
      operationId: denyApproval
      parameters:
        - $ref: "#/components/parameters/ApprovalID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Updated approval decision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "400":
          description: Invalid JSON request body.
        "404":
          description: Approval was not found.
  /v1/approvals/{id}/execute:
    post:
      summary: Execute a granted approval request once
      operationId: executeApproval
      parameters:
        - $ref: "#/components/parameters/ApprovalID"
      responses:
        "200":
          description: Execution result or blocked response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalExecuteResponse"
        "404":
          description: Approval was not found.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ApprovalID:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: approval_000001
    ToolCallID:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: tool_call_000001
  schemas:
    JSONValue:
      description: Arbitrary JSON value.
    JSONObject:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/JSONValue"
    MCPJSONRPCRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc:
          type: string
          const: "2.0"
        id:
          oneOf:
            - type: string
            - type: integer
            - type: "null"
        method:
          type: string
          example: tools/list
        params:
          $ref: "#/components/schemas/JSONObject"
    MCPJSONRPCResponse:
      type: object
      required: [jsonrpc]
      properties:
        jsonrpc:
          type: string
          const: "2.0"
        id:
          oneOf:
            - type: string
            - type: integer
            - type: "null"
        result:
          $ref: "#/components/schemas/JSONValue"
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
            message:
              type: string
...
    Connector:
      type: object
      required: [id, org_id, provider, capability, status]
      properties:
        id:
          type: string
          example: connector_000001
        org_id:
          type: string
          example: default
        name:
          type: string
          example: GitHub code_host
        provider:
          type: string
          example: github
        capability:
          type: string
          example: code_host
        config:
          $ref: "#/components/schemas/JSONObject"
        secret_ref:
          type: string
          example: env:GITHUB_TOKEN
        status:
          type: string
          enum: [configured, ready, blocked, disabled]
        source:
          type: string
          enum: [api, env, default]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ConnectorCreateRequest:
      type: object
      required: [org_id, provider, capability]
      properties:
        org_id:
          type: string
          example: default
        name:
          type: string
        provider:
          type: string
          example: github
        capability:
          type: string
          example: code_host
        config:
          $ref: "#/components/schemas/JSONObject"
        secret_ref:
          type: string
          example: env:GITHUB_TOKEN
        status:
          type: string
          enum: [configured, ready, blocked, disabled]
...
    ToolCallRequest:
      type: object
      required: [org_id, actor_user_id, agent_run_id, service_id, environment, capability, action]
      properties:
        org_id:
          type: string
        request_id:
          type: string
        actor_user_id:
          type: string
        agent_run_id:
          type: string
        service_id:
          type: string
        environment:
          type: string
        capability:
          type: string
        action:
          type: string
        arguments:
          $ref: "#/components/schemas/JSONObject"
    ToolCallResponse:
      type: object
      required: [status, risk_level]
      properties:
        status:
          type: string
          enum: [success, denied, invalid, approval_required, error]
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        provider:
          type: string
        result:
          $ref: "#/components/schemas/JSONObject"
        reason:
          type: string
        error:
          $ref: "#/components/schemas/ToolCallError"
        approval_required:
          type: boolean
        approval_request_id:
          type: string
    ToolCallError:
      type: object
      required: [retryable]
      properties:
        provider:
          type: string
        category:
          type: string
        operation:
          type: string
        status_code:
          type: integer
        attempts:
          type: integer
        retryable:
          type: boolean
        message:
          type: string
    ToolCallRecord:
      type: object
      required:
        - id
        - at
        - org_id
        - actor_user_id
        - agent_run_id
        - service_id
        - environment
        - capability
        - action
        - risk_level
        - decision
        - status
      properties:
        id:
          type: string
          example: tool_call_000001
        at:
          type: string
          format: date-time
        request_id:
          type: string
        org_id:
          type: string
        actor_user_id:
          type: string
        agent_run_id:
          type: string
        service_id:
          type: string
        environment:
          type: string
        capability:
          type: string
        action:
          type: string
        arguments:
          $ref: "#/components/schemas/JSONObject"
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        decision:
          type: string
          enum: [allowed, denied, invalid, approval_required, approved_executed]
        provider:
          type: string
        status:
          type: string
          enum: [success, denied, invalid, approval_required, error]
        reason:
          type: string
        error:
          $ref: "#/components/schemas/ToolCallError"
        approval_request_id:
          type: string
        result:
          $ref: "#/components/schemas/JSONObject"
...
    ApprovalRequest:
      type: object
      required:
        - id
        - status
        - executed
        - org_id
        - actor_user_id
        - agent_run_id
        - service_id
        - environment
        - capability
        - action
        - risk_level
        - reason
        - requested_at
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, granted, denied]
        executed:
          type: boolean
        org_id:
          type: string
        actor_user_id:
          type: string
        agent_run_id:
          type: string
        service_id:
          type: string
        environment:
          type: string
        capability:
          type: string
        action:
          type: string
        arguments:
          $ref: "#/components/schemas/JSONObject"
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        reason:
          type: string
        requested_at:
          type: string
          format: date-time
        decided_at:
          type: string
          format: date-time
        decided_by:
          type: string
        decision_note:
          type: string
        executed_at:
          type: string
          format: date-time
    ApprovalDecisionRequest:
      type: object
      required: [actor_user_id]
      properties:
        actor_user_id:
          type: string
        reason:
          type: string
    ApprovalDecisionResponse:
      type: object
      required: [status, approval]
      properties:
        status:
          type: string
          enum: [pending, granted, denied]
        approval:
          $ref: "#/components/schemas/ApprovalRequest"
    ApprovalExecuteResponse:
      type: object
      required: [status, approval]
      properties:
        status:
          type: string
          enum: [approved_executed, blocked, error]
        approval:
          $ref: "#/components/schemas/ApprovalRequest"
        tool_call:
          $ref: "#/components/schemas/ToolCallResponse"
        reason:
          type: string
...
    RiskLevel:
      type: string
      enum: [unknown, read, write_low, write_high]